move different function types to separate files
